+++ src/pages/signup/components/index.ts
@@ -1,3 +1,4 @@
 import Logo from './Logo';
+import SignUpForm from './SignUpForm';
 
-export { Logo };
+export { Logo, SignUpForm };
