+++ src/pages/signup/index.tsx
@@ -1,3 +1,3 @@
-import SignUp from './signup';
+import SignUp from './SignUp';
 
 export default SignUp;
